Reskin update ive done some good work with regard to the reskinning of the game,
Ive began to create the bones for a new object that I would like to call
levelConfig, which will be the object that houses a script that will
eventually control the many aspects of level generation, texutues,
and maybe even audio if neccessary

I envision a couple drop down menus on the levelConfig object
that will allow you to control the map size: maybe 1-5, with
1 being the smallest and 5 being the largest, and a biome selector
that skins the map accordingly. Maybe even a drop rate slider that
allows some levels to be treasure islands while most levels only
drop a moderate amount of loot.

I will keep you posted on this progress.